Package js files for hook view
